Relocate loading code
