+++ WordRot/Screens/RootScreen.swift
@@ -1,38 +1,5 @@
 import SwiftUI
 
-struct Loader {
-    static func check(pass passCallback: () -> Void, fail failCallback: () -> Void) {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        
-        let endpoint = "https://wordrot-production.herokuapp.com/api/v1/killswitch.json"
-        
-        guard
-            let killswitchUrl = URL(string: endpoint),
-            let data = try? Data(contentsOf: killswitchUrl),
-            let killswitch = try? decoder.decode(Killswitch.self, from: data)
-        else { return }
-        
-        let buildVersion = Bundle.main.object(forInfoDictionaryKey: kCFBundleVersionKey as String) as! String
-        let buildNumber = Int(buildVersion)!
-                
-        let callback = killswitch.verify(buildNumber) ? passCallback : failCallback
-        callback()
-    }
-}
-
-struct Killswitch: Codable {
-    let badBuilds: [Int]
-    let minimumBuild: Int
-    
-    func verify(_ buildNumber: Int) -> Bool {
-        let goodBuild = !badBuilds.contains(buildNumber)
-        let atLeastMinimumBuild = buildNumber >= minimumBuild
-        
-        return goodBuild && atLeastMinimumBuild
-    }
-}
-
 struct RootScreen: View {
     @State private var isLoading = true
     @State private var showPrompt = false
